Move OpenFile into it's own file
